+++ mobile/cmp/form/index.js
@@ -7,8 +7,8 @@
 
 export * from './FormField';
 export * from './Label';
-export * from './NumberInput';
-export * from './SearchInput';
-export * from './TextInput';
-export * from './TextArea';
-export * from './Select';
+export * from './inputs/NumberInput';
+export * from './inputs/SearchInput';
+export * from './inputs/TextInput';
+export * from './inputs/TextArea';
+export * from './inputs/Select';
