+++ mobile/cmp/form/inputs/Select.js
@@ -7,28 +7,19 @@
 
 import PT from 'prop-types';
 import {HoistComponent, elemFactory, LayoutSupport} from '@xh/hoist/core';
-import {castArray, isEmpty, isPlainObject, keyBy, find, assign} from 'lodash';
+import {isEmpty, isPlainObject, find, assign} from 'lodash';
 import {observable, action} from '@xh/hoist/mobx';
 import {box} from '@xh/hoist/cmp/layout';
 import {HoistInput} from '@xh/hoist/cmp/form';
 import {withDefault, throwIf} from '@xh/hoist/utils/js';
-import {
-    reactSelect,
-    reactCreatableSelect,
-    reactAsyncSelect,
-    reactAsyncCreatableSelect
-} from '@xh/hoist/kit/react-select';
+import {reactSelect} from '@xh/hoist/kit/react-select';
 
 import './Select.scss';
 
 /**
- * A managed wrapper around the React-Select combobox/dropdown component.
+ * A managed wrapper around the React-Select dropdown component.
  *
- * Supports advanced options such as:
- *      + Asynchronous queries
- *      + Multiple selection
- *      + Custom dropdown option renderers
- *      + User-created ad-hoc entries
+ * Supports custom dropdown option renderers.
  *
  * @see {@link https://react-select.com|React Select Docs}
  */
@@ -39,41 +30,14 @@ export class Select extends HoistInput {
     static propTypes = {
         ...HoistInput.propTypes,
 
-        /** True to focus the control on render. */
-        autoFocus: PT.bool,
-
-        /**
-         * Function to return a "create a new option" string prompt. Requires `allowCreate` true.
-         * Passed current query input.
-         */
-        createMessageFn: PT.func,
-
-        /** True to accept and commit input values not present in options or returned by a query. */
-        enableCreate: PT.bool,
-
-        /**
-         * True (default) to enable type-to-search keyboard input. False to disable keyboard input,
-         * showing the dropdown menu on click.
-         */
-        enableFilter: PT.bool,
-
-        /** True to allow entry/selection of multiple values - "tag picker" style. */
-        enableMulti: PT.bool,
-
         /** Field on provided options for sourcing each option's display text (default `label`). */
         labelField: PT.string,
-
-        /** Function to return loading message during an async query. Passed current query input. */
-        loadingMessageFn: PT.func,
 
         /** Placement of the dropdown menu relative to the input control. */
         menuPlacement: PT.oneOf(['auto', 'top', 'bottom']),
 
-        /** Function to return message indicating no options loaded. Passed current query input. */
+        /** Function to return message indicating no options loaded. */
         noOptionsMessageFn: PT.func,
-
-        /** True to auto-open the dropdown menu on input focus. */
-        openMenuOnFocus: PT.bool,
 
         /**
          * Preset list of options for selection. Objects must contain a `value` property; a `label`
@@ -94,12 +58,6 @@ export class Select extends HoistInput {
         placeholder: PT.string,
 
         /**
-         * Async function to return a list of options for a given query string input.
-         * Replaces the `options` prop - use one or the other.
-         */
-        queryFn: PT.func,
-
-        /**
          * Escape-hatch props passed directly to react-select. Use with care - not all props
          * in the react-select API are guaranteed to be supported by this Hoist component,
          * and providing them directly can interfere with the implementation of this class.
@@ -108,7 +66,6 @@ export class Select extends HoistInput {
 
         /** Field on provided options for sourcing each option's value (default `value`). */
         valueField: PT.string
-
     };
 
     baseClassName = 'xh-select';
@@ -117,11 +74,6 @@ export class Select extends HoistInput {
     // Maintained for (but not passed to) async select to resolve value string <> option objects.
     @observable.ref internalOptions = [];
     @action setInternalOptions(options) {this.internalOptions = options}
-
-    // Prop flags that switch core behavior.
-    get asyncMode() {return !!this.props.queryFn}
-    get creatableMode() {return !!this.props.enableCreate}
-    get multiMode() {return !!this.props.enableMulti}
 
     constructor(props) {
         super(props);
@@ -138,15 +90,14 @@ export class Select extends HoistInput {
     render() {
         const {props, renderValue} = this,
             rsProps = {
+                options: this.internalOptions,
                 value: renderValue,
 
-                autoFocus: props.autoFocus,
                 formatOptionLabel: this.formatOptionLabel,
+                isSearchable: false,
                 isDisabled: props.disabled,
-                isMulti: props.enableMulti,
                 menuPlacement: withDefault(props.menuPlacement, 'auto'),
                 noOptionsMessage: this.noOptionsMessageFn,
-                openMenuOnFocus: props.openMenuOnFocus,
                 placeholder: withDefault(props.placeholder, 'Select...'),
                 tabIndex: props.tabIndex,
 
@@ -163,33 +114,12 @@ export class Select extends HoistInput {
                 onFocus: this.onFocus
             };
 
-        if (this.asyncMode) {
-            rsProps.loadOptions = this.doQueryAsync;
-            rsProps.loadingMessage = this.loadingMessageFn;
-        } else {
-            rsProps.options = this.internalOptions;
-            rsProps.isSearchable = withDefault(props.enableFilter, true);
-        }
-
-        if (this.creatableMode) {
-            rsProps.formatCreateLabel = this.createMessageFn;
-        }
-
-        const factory = this.asyncMode ?
-            (this.creatableMode ? reactAsyncCreatableSelect : reactAsyncSelect) :
-            (this.creatableMode ? reactCreatableSelect : reactSelect);
-
         assign(rsProps, props.rsOptions);
 
         return box({
-            item: factory(rsProps),
+            item: reactSelect(rsProps),
             className: this.getClassName(),
             width: props.width,
-            onKeyDown: (e) => {
-                // Esc. can be used within the select to clear value / dismiss dropdown menu.
-                // Catch in this wrapper box - specifically to avoid dismissing dialogs.
-                if (e.key == 'Escape') e.stopPropagation();
-            },
             ...this.getLayoutProps()
         });
     }
@@ -200,17 +130,12 @@ export class Select extends HoistInput {
     //-------------------------
     onSelectChange = (opt) => {
         this.noteValueChange(opt);
-    }
+    };
 
     // Convert external value into option object(s). Options created if missing - this takes the
     // external value from the model, and we will respect that even if we don't know about it.
     // (Exception for a null value, which we will only accept if explicitly present in options.)
     toInternal(external) {
-        if (this.multiMode) {
-            if (external == null) external = [];  // avoid [null]
-            return castArray(external).map(it => this.findOption(it, !isEmpty(it)));
-        }
-
         return this.findOption(external, !isEmpty(external));
     }
 
@@ -222,9 +147,7 @@ export class Select extends HoistInput {
     }
 
     toExternal(internal) {
-        return this.multiMode ?
-            castArray(internal).map(it => it.value) :
-            isEmpty(internal) ? null : internal.value;
+        return isEmpty(internal) ? null : internal.value;
     }
 
     normalizeOptions(options) {
@@ -251,42 +174,6 @@ export class Select extends HoistInput {
             {label: src != null ? src.toString() : '-null-', value: src};
     }
 
-
-    //------------------------
-    // Async
-    //------------------------
-    doQueryAsync = (query) => {
-        return this.props
-            .queryFn(query)
-            .then(matchOpts => {
-                // Normalize query return.
-                matchOpts = this.normalizeOptions(matchOpts);
-
-                // Carry forward and add to any existing internalOpts to allow our value
-                // converters to continue all selected values in multiMode.
-                const matchesByVal = keyBy(matchOpts, 'value'),
-                    newOpts = [...matchOpts];
-
-                this.internalOptions.forEach(currOpt => {
-                    const matchOpt = matchesByVal[currOpt.value];
-                    if (!matchOpt) newOpts.push(currOpt);  // avoiding dupes
-                });
-
-                this.setInternalOptions(newOpts);
-
-                // But only return the matching options back to the combo.
-                return matchOpts;
-            });
-    }
-
-    loadingMessageFn = (params) => {
-        const {loadingMessageFn} = this.props,
-            q = params.inputValue;
-
-        return loadingMessageFn ? loadingMessageFn(q) : 'Loading...';
-    }
-
-
     //------------------------
     // Other Implementation
     //------------------------
@@ -306,21 +193,13 @@ export class Select extends HoistInput {
         // Always display the standard label string in the value container (context == 'value').
         // If we need to expose customization here, we could consider a dedicated prop.
         return (optionRenderer && params.context == 'menu') ? optionRenderer(opt) : opt.label;
-    }
+    };
 
-    noOptionsMessageFn = (params) => {
-        const {noOptionsMessageFn} = this.props,
-            q = params.inputValue;
-
-        if (noOptionsMessageFn) return noOptionsMessageFn(q);
-        if (q) return 'No matches found.';
-        return this.asyncMode ? 'Type to search...' : '';
-    }
-
-    createMessageFn = (q) => {
-        const {createMessageFn} = this.props;
-        return createMessageFn ? createMessageFn(q) : `Create "${q}"`;
-    }
+    noOptionsMessageFn = () => {
+        const {noOptionsMessageFn} = this.props;
+        if (noOptionsMessageFn) return noOptionsMessageFn();
+        return 'No options found.';
+    };
 
     getOrCreatePortalDiv() {
         let portal = document.getElementById('xh-select-input-portal');
@@ -335,4 +214,5 @@ export class Select extends HoistInput {
     }
 
 }
+
 export const select = elemFactory(Select);
